requirements.txt
kfp==2.*
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
pandas
scikit-learn
joblib
yaml